Vendorize icon assets and ensure they are referenced when published
This copies the actual icon SVGs we use into the repository, so that
we can ship them with the package without needing to ship every single
icon offered by Shoelace.

It also creates a new utility function so that the asset path is set
correctly in both local dev and when published.